refactor: Remove Folium map stack for Leaflet-only production maps
Refs: #222
