+++ benchmarks/map_perf/stage_ceilings.json
@@ -5,11 +5,11 @@
   "max_elapsed_ms": {
     "dataset.load": 180000,
     "prep.map_context_prepare": 180000,
-    "prep.build_species_overlay_map": 180000,
     "prep.cached_family_map_bundle": 180000,
     "prep.family_map_composition_with_pins": 180000,
-    "prep.folium_map_to_html_bytes": 240000,
-    "prep.map_iframe_embed": 120000,
+    "map.all_locations_leaflet.component_embed": 120000,
+    "map.all_locations_leaflet.payload": 240000,
+    "map.lifer_leaflet.component_embed": 120000,
     "taxonomy.cached_species_url_fn": 60000
   }
 }
