+++ explorer/app/streamlit/map_working.py
@@ -118,34 +118,3 @@ def streamlit_working_set_and_status(
         return None, "Date filter: invalid range"
     return ws, f"Date filter: {start_s} to {end_s}"
 
-
-def folium_map_to_html_bytes(m: Any) -> bytes:
-    """Serialize a Folium map to UTF-8 HTML bytes for download or iframe ``srcdoc``.
-
-    ``branca`` full-document ``render()`` mutates the in-memory map; callers that keep a long-lived
-    :class:`folium.Map` in session cache must pass a **copy** of the map here (not the cached
-    object).
-    """
-    root = m.get_root()
-    html = root.render()
-    if isinstance(html, bytes):
-        return html
-    return str(html).encode("utf-8")
-
-
-def embed_folium_html_bytes_iframe(
-    html_bytes: bytes,
-    *,
-    height: int,
-    scrolling: bool = True,
-) -> None:
-    """Render pre-serialized Folium full-page HTML in the app via :func:`streamlit.components.v1.html`.
-
-    Streamlit’s HTML iframe (``srcdoc=…``) matches what ``st_folium`` ultimately ships to the
-    client; the same *bytes* used for **Export map HTML** can be passed here for one render per
-    run (or zero when ``html_bytes`` is served from the map HTML cache).
-    """
-    import streamlit.components.v1 as components
-
-    s = html_bytes.decode("utf-8")
-    components.html(s, width=None, height=int(height), scrolling=scrolling)
