+++ explorer/core/all_locations_experimental_marker_style.py
@@ -1,11 +1,11 @@
-"""Resolved Folium-equivalent All locations pin styling for the Leaflet map."""
+"""Resolved All locations pin styling for the Leaflet map component."""
 
 from __future__ import annotations
 
 from typing import Any
 
 from explorer.app.streamlit.defaults import active_map_marker_colour_scheme
-from explorer.core.map_overlay_visit_map import (
+from explorer.core.map_leaflet_viewport import (
     _all_locations_marker_params_from_scheme,
     all_locations_cluster_icon_style_payload,
 )
@@ -25,11 +25,11 @@ def circle_marker_style_for_all_locations_map(colour_scheme_index: int) -> dict[
 
 
 def cluster_icon_style_for_all_locations_map(colour_scheme_index: int) -> dict[str, Any] | None:
-    """MarkerCluster ``iconCreateFunction`` inputs for the Leaflet component (Folium parity)."""
+    """MarkerCluster tier colours for the Leaflet component."""
     sch = active_map_marker_colour_scheme(int(colour_scheme_index))
     return all_locations_cluster_icon_style_payload(sch)
 
 
 def experimental_default_scheme_circle_marker_props() -> dict[str, Any]:
-    """Preset **1** (Eucalypt) — kept for spike-era callers; production uses :func:`circle_marker_style_for_all_locations_map`."""
+    """Preset **1** (Eucalypt) — kept for spike-era callers."""
     return circle_marker_style_for_all_locations_map(1)
