+++ explorer/core/lifer_locations_geojson.py
@@ -1,6 +1,6 @@
 """GeoJSON + revision for the **Lifer locations** Leaflet component.
 
-Mirrors :func:`~explorer.core.map_overlay_lifer_map.build_lifer_overlay_map` data without Folium.
+Mirrors :func:`~explorer.core.map_leaflet_viewport.build_lifer_overlay_map` data without Folium.
 Structured ``lifer_popup_v1`` feeds the iframe TS template (parity with ``format_lifer_popup_lines``).
 """
